Include safety checks in _calc_dist2root for missing branch lengths
If TreeTime is handed a Newick tree with just a topology, the _calc_dist2root
function will fail reporting
TypeError: unsupported operand type(s) for +: 'float' and 'NoneType'
This resulted from the branch_length attr existing but being set to None. With
this fix TreeTime now works with just an input topology.