Return the total likelyhood of the uncompressed tree from ancestral_likelihood()
Previously the function returned array of likelyhoods corresponding to
the columns of the compressed alignment. Tests updated to match. This
change was motivated in part by the incompatibility of
test_seq_joint_reconstruction_ss_correct() with the old version of
ancestral_likelyhood().

Comparing likelihoods of uncompressed sequences in tests for joint
reconstruction