Add missing entry to BLOSUM45 matrix
Adds a missing entry, 0, to the last row of the BLOSUM45 matrix so all 20 rows
contain the expected 20 values each. Resolves a visible deprecation warning from
numpy 1.19.0 that reports automatic conversion of ragged nested arrays as
`VisibleDeprecationWarning: Creating an ndarray from ragged nested
sequences (which is a list-or-tuple of lists-or-tuples-or ndarrays with
different lengths or shapes) is deprecated. If you meant to do this, you must
specify 'dtype=object' when creating the ndarray`.

For more information about the deprecation in numpy see the release notes for
version 1.19.0 [1].

[1] https://numpy.org/doc/1.19/release/1.19.0-notes.html#deprecate-automatic-dtype-object-for-ragged-input