[vcf] Improve VCF parsing
The VCF 4.2/4.3 spec is precise about the format of meta-information /
header-line / data-line syntax. We assert that the file contents matches
what we expect and raise errors if not. This allows code simplifications.

I've reordered the if/else block to match how the file appears, not the
previous approach of placing the data lines first as they're the "most
common so first in 'if-list'". This is much more readable to me!