[vcf] write unchanged ref genotype as 0
The VCF spec uses a genotype of '0' to mean that the reference base(s)
is unchanged for this sample. The previous '.' means an uncalled allele,
in v4.2 "a call cannot be made for a sample at a given locus" and in
v4.3 "a call cannot be made for a sample at a given locus".