[vcf] use metadata to influence write format
This is only used if provided, otherwise we fall back to the default
ploidy of diploid (unchanged) and chromosome name of "1" (changed from
the previously hardcoded "MTB_anc"). Note that the spec states that
no-call genotypes should be encoded to reflect the ploidy (e.g. "./."
for diploid and "." only for haploid) which is a change from previous
behaviour.

Closes #262