fix: avoid empty entries in `non_consensus` field
In the `non_consensus` dict after we pop elements from the `states` sets of some of the positions, positions with empty state sets remained.

Let's fix that by reiterating the dicts and retaining only positions with non-empty sets.

We might check whether this is needed, or we could simply cleanup the whole dict at the end.