fix: crash in fitch recostruction
Our tests conveniently tested reconstruction for a tree having only 2 levels or internal nodes, meaning that during reconstruction parent sequences other than for root node not need to be accessed.

However, with deeper trees the implementation crashed, because once we reach 3rd level, there is no parent sequence to read from.

Here I memorize the reconstructed sequence of the node, meaning that all reconstructed sequences are stored until the end of the program. In order to reduce memory footprint, we might use biphasic traversal - the second phase would clean up the sequences when they are no longer needed (when the node becomes a grandparent of the currently visited node generation). But would be a bandaid on top of generally poor software design. There might be a better way to implement the reconstruction without storing sequences.

The bug was additionally hidden by the "partially initialized object" anti-pattern, widespread in the codebase now. It's hard to reason about validity of objects, because any part of any object can be uninitialized at any time. So it is difficult to ensure or even check that the correct fields are initialized when they are needed. This needs to be redesigned.