feat(clock): add new algos for edge splitting during reroot
This rearranges code for `find_best_root()` and `find_best_split()` to allow for configurable optimization methods in `find_best_split()`.

The old grid search code is rearranged, but preserved. Added Brent search and Golden Section search algos.

The cost function is the same for all three.

Currently the choice is hardcoded to be grid search but it can be made configurable, e.g. using CLI args - the interface is to be defined and implemented.